This is fixing a problem in 'TestForEachRight' that appears because it incorrectly relies on the underlying implementation of Map in Golang
The ordering is not guaranteed, but rather the slice output format is randomized strictly to disallow reliying on it. This happens at least for all Goversions >=1.7.

Check that numbers are in the output without asserting the order.

This is also the reason why unit tests on master fail every now and then, and made worse by running tests on 1.8 and 1.9!

Provoke the problem by running the test quite repeatadly until it fails:

'(set -e && for i in {1..100}; do go test; done)'